refactor(test): Move ClassWithSwitchExample.java to testclasses (#2158)
